// src/frontend/src/common/graphql/types.ts
export type CreateChatInput = {
  userId?: string | null,
  sessionId: string,
  human?: string | null,
  assistant?: string | null,
  trace?: string | null,
  traceMetadata?: string | null,
  expiration?: number | null,
  id?: string | null,
  sessionChatsId?: string | null,
};

export type ModelChatConditionInput = {
  userId?: ModelIDInput | null,
  sessionId?: ModelStringInput | null,
  human?: ModelStringInput | null,
  assistant?: ModelStringInput | null,
  trace?: ModelStringInput | null,
  traceMetadata?: ModelStringInput | null,
  expiration?: ModelIntInput | null,
  and?: Array< ModelChatConditionInput | null > | null,
  or?: Array< ModelChatConditionInput | null > | null,
  not?: ModelChatConditionInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  sessionChatsId?: ModelIDInput | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type Chat = {
  __typename: "Chat",
  userId?: string | null,
  sessionId: string,
  human?: string | null,
  assistant?: string | null,
  trace?: string | null,
  traceMetadata?: string | null,
  expiration?: number | null,
  id: string,
  createdAt: string,
  updatedAt: string,
  sessionChatsId?: string | null,
};

export type UpdateChatInput = {
  userId?: string | null,
  sessionId?: string | null,
  human?: string | null,
  assistant?: string | null,
  trace?: string | null,
  traceMetadata?: string | null,
  expiration?: number | null,
  id: string,
  sessionChatsId?: string | null,
};

export type DeleteChatInput = {
  id: string,
};

export type CreateSessionInput = {
  userId?: string | null,
  expiration?: number | null,
  id?: string | null,
};

export type ModelSessionConditionInput = {
  userId?: ModelIDInput | null,
  expiration?: ModelIntInput | null,
  and?: Array< ModelSessionConditionInput | null > | null,
  or?: Array< ModelSessionConditionInput | null > | null,
  not?: ModelSessionConditionInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
};

export type Session = {
  __typename: "Session",
  userId?: string | null,
  chats?: ModelChatConnection | null,
  expiration?: number | null,
  id: string,
  createdAt: string,
  updatedAt: string,
};

export type ModelChatConnection = {
  __typename: "ModelChatConnection",
  items:  Array<Chat | null >,
  nextToken?: string | null,
};

export type UpdateSessionInput = {
  userId?: string | null,
  expiration?: number | null,
  id: string,
};

export type DeleteSessionInput = {
  id: string,
};
// ...
export type SendChatMutationVariables = {
  sessionId: string,
  human: string,
  sessionAttributes?: string | null,
};

export type SendChatMutation = {
  sendChat?: string | null,
};

export type CreateChatMutationVariables = {
  input: CreateChatInput,
  condition?: ModelChatConditionInput | null,
};

export type CreateChatMutation = {
  createChat?:  {
    __typename: "Chat",
    userId?: string | null,
    sessionId: string,
    human?: string | null,
    assistant?: string | null,
    trace?: string | null,
    traceMetadata?: string | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
    sessionChatsId?: string | null,
  } | null,
};

export type UpdateChatMutationVariables = {
  input: UpdateChatInput,
  condition?: ModelChatConditionInput | null,
};

export type UpdateChatMutation = {
  updateChat?:  {
    __typename: "Chat",
    userId?: string | null,
    sessionId: string,
    human?: string | null,
    assistant?: string | null,
    trace?: string | null,
    traceMetadata?: string | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
    sessionChatsId?: string | null,
  } | null,
};

export type DeleteChatMutationVariables = {
  input: DeleteChatInput,
  condition?: ModelChatConditionInput | null,
};

export type DeleteChatMutation = {
  deleteChat?:  {
    __typename: "Chat",
    userId?: string | null,
    sessionId: string,
    human?: string | null,
    assistant?: string | null,
    trace?: string | null,
    traceMetadata?: string | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
    sessionChatsId?: string | null,
  } | null,
};

export type CreateSessionMutationVariables = {
  input: CreateSessionInput,
  condition?: ModelSessionConditionInput | null,
};

export type CreateSessionMutation = {
  createSession?:  {
    __typename: "Session",
    userId?: string | null,
    chats?:  {
      __typename: "ModelChatConnection",
      nextToken?: string | null,
    } | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type UpdateSessionMutationVariables = {
  input: UpdateSessionInput,
  condition?: ModelSessionConditionInput | null,
};

export type UpdateSessionMutation = {
  updateSession?:  {
    __typename: "Session",
    userId?: string | null,
    chats?:  {
      __typename: "ModelChatConnection",
      nextToken?: string | null,
    } | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type DeleteSessionMutationVariables = {
  input: DeleteSessionInput,
  condition?: ModelSessionConditionInput | null,
};

export type DeleteSessionMutation = {
  deleteSession?:  {
    __typename: "Session",
    userId?: string | null,
    chats?:  {
      __typename: "ModelChatConnection",
      nextToken?: string | null,
    } | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
  } | null,
};